requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile pyproject.toml
#
anyio==3.6.1
    # via starlette
click==8.1.3
    # via distributed
cloudpickle==2.1.0
    # via
    #   dask
    #   distributed
dask==2022.5.2
    # via
    #   dask-jobqueue
    #   distributed
    #   juqueue (pyproject.toml)
dask-jobqueue==0.7.3
    # via juqueue (pyproject.toml)
distributed==2022.5.2
    # via dask-jobqueue
fastapi==0.78.0
    # via juqueue (pyproject.toml)
filelock==3.7.1
    # via juqueue (pyproject.toml)
fsspec==2022.5.0
    # via dask
h11==0.13.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
heapdict==1.0.1
    # via zict
hpack==4.0.0
    # via h2
hypercorn==0.13.2
    # via juqueue (pyproject.toml)
hyperframe==6.0.1
    # via h2
idna==3.3
    # via anyio
jinja2==3.1.2
    # via distributed
locket==1.0.0
    # via
    #   distributed
    #   partd
loguru==0.6.0
    # via juqueue (pyproject.toml)
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.4
    # via distributed
nest-asyncio==1.5.5
    # via juqueue (pyproject.toml)
packaging==21.3
    # via
    #   dask
    #   distributed
partd==1.2.0
    # via dask
priority==2.0.0
    # via hypercorn
psutil==5.9.1
    # via distributed
pydantic==1.9.1
    # via
    #   fastapi
    #   juqueue (pyproject.toml)
pyparsing==3.0.9
    # via packaging
pyyaml==6.0
    # via
    #   dask
    #   distributed
    #   juqueue (pyproject.toml)
sniffio==1.2.0
    # via anyio
sortedcontainers==2.4.0
    # via distributed
starlette==0.19.1
    # via fastapi
tblib==1.7.0
    # via distributed
toml==0.10.2
    # via hypercorn
toolz==0.11.2
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via
    #   distributed
    #   juqueue (pyproject.toml)
typing-extensions==4.2.0
    # via pydantic
urllib3==1.26.9
    # via distributed
uvloop==0.16.0
    # via juqueue (pyproject.toml)
wsproto==1.1.0
    # via hypercorn
zict==2.2.0
    # via distributed

aioprocessing[dill]
pytest
pytest-asyncio
python-socketio[asyncio_client]
aiohttp[speedups]
websocket-client
requests
